+++ src/commonMain/kotlin/com/petersamokhin/notionsdk/data/mapper/DatabasePageMapper.kt
@@ -6,5 +6,6 @@ import com.petersamokhin.notionsdk.data.model.result.NotionDatabaseRow
 
 internal fun PageObject.toDomain(): NotionDatabaseRow =
     NotionDatabaseRow(
+        id = id,
         columns = properties.mapValues { (key, value) -> NotionDatabaseColumn(key, value.toDomain()) }
     )