+++ src/commonMain/kotlin/com/petersamokhin/notionsdk/data/model/internal/obj/User.kt
@@ -1,33 +1,43 @@
-package com.petersamokhin.notionsdk.data.model.result
+package com.petersamokhin.notionsdk.data.model.internal.obj
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-public sealed class NotionUser {
-    public abstract val id: String
-    public abstract val name: String?
+internal sealed class User {
+    abstract val id: String
+    abstract val name: String?
 
     @SerialName("avatar_url")
-    public abstract val avatarUrl: String?
+    abstract val avatarUrl: String?
 
     @Serializable
     @SerialName("person")
-    public data class Person(
+    data class Person(
         override val id: String,
         override val name: String? = null,
         @SerialName("avatar_url")
         override val avatarUrl: String? = null,
+        val person: Value,
+    ) : User() {
 
-        public val email: String,
-    ) : NotionUser()
+        @Serializable
+        data class Value(
+            val email: String,
+        )
+    }
 
     @Serializable
     @SerialName("bot")
-    public data class Bot(
+    data class Bot(
         override val id: String,
         override val name: String? = null,
         @SerialName("avatar_url")
         override val avatarUrl: String? = null,
-    ) : NotionUser()
+        val bot: Bot? = null,
+    ) : User() {
+
+        @Serializable
+        object Bot
+    }
 }